Changed ordered rules to use `Comparator<T>`
This commit changes the way the compatibility and disambiguation rules work, by using
a `Comparator<T>` instead of relying on the attribute to implement `Comparable`. While
making it more verbose in case the attribute implements `Comparable`, it saves us from
generics madness, and supports more cases (attributes are not required to be comparable).